Updating to fix some "minor errors"

+++ Source/Utilities.cpp
@@ -807,7 +807,7 @@ namespace DiscordCoreAPI {
 		}
 		newBuffer[1] = encodedBytes;
 		newBuffer[2] = 0;
-		this->writeString(newBuffer, 1 + 2 + encodedBytes);
+		this->writeString(newBuffer, 1ull + 2ull + static_cast<size_t>(encodedBytes));
 	}
 
 	void Jsonifier::appendNewFloatExt(const double FloatValue) {
